requirements.txt
streamlit
gspread
google-auth
fpdf